Move utility files into utils submodule
