+++ gwdetchar/utils/tests/test_daq.py
@@ -73,9 +73,9 @@ def test_ligo_accum_overflow_channel():
         'X1:FEC-4_ACCUM_OVERFLOW')
 
 
-@mock.patch('gwdetchar.daq.find_urls')
-@mock.patch('gwdetchar.daq.get_channel_names')
-@mock.patch('gwdetchar.daq._ligo_model_overflow_channels_nds')
+@mock.patch('gwdetchar.utils.daq.find_urls')
+@mock.patch('gwdetchar.utils.daq.get_channel_names')
+@mock.patch('gwdetchar.utils.daq._ligo_model_overflow_channels_nds')
 def test_ligo_model_overflow_channels(nds, get_names, find_frames):
     get_names.return_value = CHANNELS
